Extend CommandHandler to handle up to 11 arguments
This patch fixes `System.CommandLine.Invocation.CommandHandler.Create`
so that it can handle actions with up to 11 arguments. Current limit of
7 arguments is not enough for a more complex applications.

While users can alternatively bind arguments to arbitrary number of
properties, this requires refactoring the code once the limit is
hit. Additionally, some users are not aware of the feature and could
hit the wall unintentionally. With this patch, the arbitrary, but
reasonable limit of 11 arguments should cover most of the use cases
"in the wild".

Fix #983.